Resolves DynamicsValue/fake-xrm-easy#144
